satyam/satyam/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-bootstrap-form==3.4
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
django-widget-tweaks==1.4.8
phonenumbers==8.12.24
Pillow==8.0.1
PyMySQL==0.10.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
